ArtUpdate
Animations for walking in all directions is done. The up and down animations are 2 different child objects. in the script just turn them active once the player goes up or down. the for the left animation just set the scale to negative to flip it. In the animator there is the idle and walking animation. GL